refactor collection data source
